requirements.txt
matplotlib>=3.1.1
scikit-learn>=0.22.1
pynverse~=0.1.4.4
numpy~=1.19.5
scipy~=1.7.1
pillow~=8.3.2
scikit-image~=0.18.3
opencv-python~=4.1.2
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.9.0+cpu
torchvision==0.10.0+cpu